feat: Authentication 관련 파일 auth 디렉토리로 이동
